requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangGraph and AI
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# Instagram/Meta Integration
requests>=2.31.0
aiohttp>=3.9.0
cryptography>=41.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0          # PostgreSQL async driver
aiosqlite>=0.19.0        # SQLite async driver
redis>=5.0.0
aiofiles>=23.2.0

# MCP (Model Context Protocol)
mcp>=1.12.0

# Razorpay Integration
razorpay>=1.4.0

# Voice Services (Future Phase)
# deepgram-sdk>=3.2.0
# whisper>=1.1.0

# Security and Compliance
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0
langfuse>=2.50.0  # LLM observability and tracing
# opentelemetry-api>=1.21.0
# opentelemetry-sdk>=1.21.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development Tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Google Sheets Integration
gspread>=6.0.0
google-auth>=2.0.0

# Environment Management
python-dotenv>=1.0.0
